requirements.txt
typer[all]
pyyaml
requests